WIP: Registering timestamped edge 32bit for reading binary graph partition.
